ev-scheduling/requirements.txt
# ECE 285 Project — EV Charging Schedule Assistant
# Install from project root: pip install -r requirements.txt

# Data (acnportal is in ./acnportal; install locally or add -e ./acnportal)
-e ./acnportal

# Optimization
cvxpy>=1.4.0

# LLM / agent
openai>=1.0.0

# Numerics and viz
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0

# Web GUI
fastapi>=0.110.0
uvicorn>=0.29.0
